refactor(editor): Nest sanitization constants under constants/ in @n8n/composables (no-changelog)
Per review feedback, the sanitization constants now live at
@n8n/composables/src/constants/sanitization.ts (git mv, history preserved).
Repoints htmlUtils's import and the editor-ui @/app/constants/sanitization
shim to the @n8n/composables/constants/sanitization subpath. The package's
wildcard ./* export resolves the nested dist path, so no export-map change
is needed. No behavior change.
