+++ packages/frontend/@n8n/composables/src/constants/sanitization.ts
